Fix prefjson::setPreferences() doesn't actually save.
