--- src/core/tracker/tracker.pri
@@ -1,9 +1,0 @@
-INCLUDEPATH += $$PWD
-
-HEADERS += \
-    $$PWD/qtracker.h \
-    $$PWD/trackerannouncerequest.h \
-    $$PWD/qpeer.h
-
-SOURCES += \
-    $$PWD/qtracker.cpp
